+++ docs/suspensive.org/src/content/en/docs/react-query/_meta.tsx
@@ -3,15 +3,10 @@ import type { MetaRecord } from 'nextra'
 export default {
   motivation: { title: 'Motivation' },
   installation: { title: 'Installation' },
-  migration: { title: 'Migration Guide' },
   '--- API Reference': {
     type: 'separator',
     title: 'API Reference',
   },
-  createGetQueryClient: { title: 'createGetQueryClient' },
-  mutationOptions: { title: 'mutationOptions' },
-  usePrefetchQuery: { title: 'usePrefetchQuery' },
-  usePrefetchInfiniteQuery: { title: 'usePrefetchInfiniteQuery' },
   SuspenseQuery: { title: '<SuspenseQuery/>' },
   SuspenseQueries: { title: '<SuspenseQueries/>' },
   SuspenseInfiniteQuery: { title: '<SuspenseInfiniteQuery/>' },
@@ -23,6 +18,10 @@ export default {
   QueriesHydration: { title: '<QueriesHydration/>' },
   QueryClientConsumer: { title: '<QueryClientConsumer/>' },
   IsFetching: { title: '<IsFetching/>' },
+  createGetQueryClient: { title: 'createGetQueryClient' },
+  mutationOptions: { title: 'mutationOptions' },
+  usePrefetchQuery: { title: 'usePrefetchQuery' },
+  usePrefetchInfiniteQuery: { title: 'usePrefetchInfiniteQuery' },
   '--- Deprecated': {
     type: 'separator',
     title: 'Deprecated',
@@ -32,4 +31,9 @@ export default {
   useSuspenseQuery: { title: 'useSuspenseQuery' },
   useSuspenseQueries: { title: 'useSuspenseQueries' },
   useSuspenseInfiniteQuery: { title: 'useSuspenseInfiniteQuery' },
+  '--- More': {
+    type: 'separator',
+    title: 'More',
+  },
+  migration: { title: 'Migration Guide' },
 } satisfies MetaRecord