+++ docs/suspensive.org/src/content/ko/_meta.tsx
@@ -12,7 +12,17 @@ export default {
   },
   docs: {
     type: 'page',
-    title: '문서보기',
+    title: '문서',
+  },
+  'getting-started': {
+    type: 'page',
+    title: '시작하기',
+    href: '/docs/react/getting-started',
+  },
+  compare: {
+    type: 'page',
+    title: '비교',
+    href: '/docs/react/comparison',
   },
   versions: {
     type: 'menu',